Extract base class OpenApiCodeGenerator
